+++ resources/lang/ca/battles.php
@@ -41,5 +41,5 @@ return [
     'monster_empty' => 'No tens monstres, crea un per a jugar en aquesta secció',
     'monster_hp_loss' => 'Pèrdua de punts de vida del monstre',
     'defeat' => 'us ha vençut!',
-    '' => '',
+    'show_hide_img' => 'Mostra / amaga imatge',
 ];